[Refactor] 3 Again Satisfy clang-tidy

+++ tst/FullJitterTest.cpp
@@ -27,7 +27,7 @@ TEST(FullJitterTest, ZeroInputAlwaysReturnsZero) {
 TEST(FullJitterTest, DistributionIsUniformEnough) {
     FullJitter jitter;
     auto input = std::chrono::microseconds(10);
-    std::set<int> seen;
+    std::set<long> seen;
     for (int i = 0; i < 1000; ++i) {
         auto result = jitter.jitter(input);
         seen.insert(result.count());
